requirements.txt
pandas
numpy
scikit-learn
lightgbm
xgboost
yfinance
requests
fastapi
uvicorn
python-dotenv
pytest
loguru
azure-identity
azure-keyvault-secrets
azure-storage-blob
openai
transformers
beautifulsoup4
lxml
sqlalchemy
pydantic
joblib
shap
